# design/openapi.yaml
...
openapi: 3.0.3
info:
  title: DriveDeal — Reverse Car Marketplace API
  version: 1.0.0
  description: |-
    Backend API for DriveDeal: a reverse car marketplace where buyers post the
    car they want and dealers compete on price for it.

    Handles:
    - Buyer flows: post a car request, watch the live quote leaderboard, accept a
      quote, track the deal, chat with the dealer.
    - Dealer flows: account with manual approval, geo-filtered request feed, quote
      submission and revision, deal status.
    - Support flows: ticket queue, dealer/agent verification approval, roster.
    - AI flows (buyer only): a streaming advisory agent and a compare agent,
      grounded in the database plus a Crawl4AI web search.

    Conventions:
    - All JSON routes are under /api/v1 (this server URL). Exceptions: GET /health
      (raw health object) and the /api/v1/ws WebSocket.
    - **Success responses are envelope-free.** The payload is the resource itself.
      There is no {"success": true, "data": ...} wrapper — the frontend's mock
      client already returns bare objects and that shape is the contract
      (solution-frontend.md §2.1, solution-backend.md §0.2).
    - **Every non-2xx JSON response uses the error envelope**, and nothing else:
        {"error": {"code", "message", "details", "request_id"}}
      `code` is a stable machine-readable string the client branches on.
      `message` is for humans and may be reworded; never parse it.
...
  contact:
...
servers:
...
security:
- bearer_auth: []
paths:
...
  /auth/signup/buyer:
    post:
      summary: Buyer signup
      description: |-
        Creates a `profiles` row (role=buyer) and a `users` row with an Argon2id
        password hash, in one transaction. Returns access + refresh tokens.

        The password is never logged. Response is 201.
      operationId: signupBuyer
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyerSignupRequest'
            examples:
              valid:
                value:
                  full_name: Adithyaa Anand
                  email: adithyaa.ma@gmail.com
                  phone: '+14695550142'
                  password: sunroof2024
                  terms_accepted: true
                  terms_version: '2026-09-01'
      responses:
        '201':
          description: Buyer created and signed in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
              examples:
                created:
                  value:
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    token_type: bearer
                    expires_in: 1800
                    profile:
                      $ref: '#/components/examples/BuyerProfile'
        '409':
          $ref: '#/components/responses/conflict'
        '422':
          $ref: '#/components/responses/validation_error'
        '429':
          $ref: '#/components/responses/rate_limited'
  /auth/signup/dealer:
    post:
      summary: Dealer signup — pending manual approval
      description: |-
        Creates the profile (role=dealer), a `users` row with
        `is_active = false`, and a `support_verifications` row with
        `category = 'dealer'`, `status = 'pending'` and
        `ticket_id = 'DV<epoch>'`.

        **No tokens are issued.** The account is inert until a support agent
        approves the verification, which flips `users.is_active = true`. The
        client renders a pending state, not a login redirect.
      operationId: signupDealer
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealerSignupRequest'
            examples:
              valid:
                value:
                  full_name: Rahul Sharma
                  email: rahul01@gmail.com
                  phone: '+14695550143'
                  password: dealer2024x
                  dealership_name: Navee Motors
                  branch_name: Frisco
                  dealer_license: US030202
                  state_id: 0e1f2a3b-4c5d-4e6f-8a9b-0c1d2e3f4a5b
                  address: 4100 Legacy Dr, Frisco, TX
                  website: https://naveemotors.example
                  brand_ids:
                  - 1a2b3c4d-5e6f-4a7b-8c9d-0e1f2a3b4c5d
                  terms_accepted: true
                  terms_version: '2026-09-01'
      responses:
        '201':
          description: Dealer created, pending approval (no tokens)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingApproval'
              examples:
                pending:
                  value:
                    pending: true
                    role: dealer
                    email: rahul01@gmail.com
                    verification_id: 4cac1879-8502-42ba-8140-d321aba9b77
                    ticket_id: DV1788793917
                    message: Your dealer account is awaiting support review — you'll get an email once
                      it's approved.
        '409':
          $ref: '#/components/responses/conflict'
        '422':
          $ref: '#/components/responses/validation_error'
        '429':
          $ref: '#/components/responses/rate_limited'
  /auth/signup/support:
    post:
      summary: Support agent signup — pending manual approval
      description: |-
        Support team accounts require manual approval.

        Creates the profile (role=support), a `users` row with
        `is_active = false`, and a `support_verifications` row with
        `category = 'agent'`, `status = 'pending'`,
        `ticket_id = 'SA<epoch>'`.

        **No tokens are issued.** Approving an `agent` verification requires the
        `admin` role — a support agent can never approve a peer
        (solution-backend.md §9.5).

        This route is **not linked from public navigation** and is reachable by
        direct URL only. The UI hiding is cosmetic; the real control is that the
        account is inert until approved.
      operationId: signupSupport
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportSignupRequest'
            examples:
              valid:
                value:
                  full_name: Support03
                  phone: '+14695550144'
                  address: 2200 Ross Ave, Dallas, TX
                  state_id: 0e1f2a3b-4c5d-4e6f-8a9b-0c1d2e3f4a5b
                  extra_info: Former dealership service manager, 6 years.
                  email: support03@gmail.com
                  password: support2024
                  terms_accepted: true
                  terms_version: '2026-09-01'
      responses:
        '201':
          description: Agent created, pending approval (no tokens)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingApproval'
              examples:
                pending:
                  value:
                    pending: true
                    role: support
                    email: support03@gmail.com
                    verification_id: b1e876b4-da1f-41e6-9eaf-1955cf13af4f
                    ticket_id: SA97379
                    message: Your support account is pending approval. Our team reviews every request
                      — you'll get an email once it's approved.
        '409':
          $ref: '#/components/responses/conflict'
        '422':
          $ref: '#/components/responses/validation_error'
        '429':
          $ref: '#/components/responses/rate_limited'
  /auth/login:
    post:
      summary: Login
      description: |-
        One login for all roles — the role is read from the session, never
        selected by the user.

        - Wrong password, or an unknown email → 401 INVALID_CREDENTIALS. A dummy
          Argon2id verify runs on unknown emails so response time does not
          enumerate accounts.
        - `users.is_active = false` with a pending verification → 403
          DEALER_PENDING_REVIEW (or SUPPORT_PENDING_REVIEW). This one deliberate
          distinction is the only difference in the response.
        - A suspended account is 403 with code FORBIDDEN_ROLE.
      operationId: login
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              buyer:
                value:
                  email: rahul01@gmail.com
                  password: sunroof2024
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
              examples:
                ok:
                  value:
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    token_type: bearer
                    expires_in: 1800
                    profile:
                      $ref: '#/components/examples/BuyerProfile'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/pending_review'
        '422':
          $ref: '#/components/responses/validation_error'
        '429':
          $ref: '#/components/responses/rate_limited'
  /auth/refresh:
    post:
      summary: Refresh the access token
      description: |-
        Rotates the refresh token's `jti`. Presenting a rotated (already used)
        token revokes the whole family — replay detection, not a silent reissue.
      operationId: refreshToken
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/unauthorized'
  /auth/logout:
    post:
      summary: Revoke the current session
      operationId: logout
      tags:
      - Auth
      responses:
        '200':
          description: Session revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Signed out
        '401':
          $ref: '#/components/responses/unauthorized'
  /auth/forgot-password:
    post:
      summary: Request a password reset
      description: |-
        **Always 202**, whether or not the email exists. Returning 202 for both
        is what prevents account enumeration.
      operationId: forgotPassword
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '202':
          description: Accepted regardless of whether the email exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: If that email has an account, a reset link is on its way.
        '429':
          $ref: '#/components/responses/rate_limited'
  /auth/reset-password:
    post:
      summary: Complete a password reset
      description: Hashes with Argon2id and revokes every existing session for the profile.
      operationId: resetPassword
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - token
              - password
              properties:
                token:
                  type: string
                  description: Single-use reset token from the email link
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password updated. Sign in with your new password.
        '400':
          $ref: '#/components/responses/invalid_token'
        '422':
          $ref: '#/components/responses/validation_error'
  /auth/me:
    get:
      summary: Current session profile
      description: |-
        Profile plus role plus verification status. The client routes on `role`;
        there is no role picker in the UI.
      operationId: getCurrentUser
      tags:
      - Auth
      responses:
        '200':
          description: Current profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
              examples:
                buyer:
                  value:
                    $ref: '#/components/examples/MeBuyer'
        '401':
          $ref: '#/components/responses/unauthorized'
...
components:
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |-
        RS256 access token from /auth/login or /auth/refresh.

        Subject = `profiles.id` (NOT the `users.id`, which is a different
        UUID and does not appear anywhere in the request layer). Roles are
        read from the `profiles` row on every request, so a role change or a
        suspension takes effect on the next call rather than at token expiry.

        For the WebSocket, use a one-shot ticket from /chats/ws-ticket
        instead — a browser cannot set this header on the handshake.
  responses:
    unauthorized:
      description: |-
        Missing, malformed, expired, or revoked token. `code`
        distinguishes: `UNAUTHORIZED`, `TOKEN_EXPIRED`, `TOKEN_REVOKED`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            expired:
              value:
                error:
                  code: TOKEN_EXPIRED
                  message: Your session has expired. Sign in again.
                  details: null
                  request_id: 3f2a1b0c-9d8e-4f7a-8b6c-5d4e3f2a1b0c
...
    pending_review:
      description: |-
        The account exists but is inactive pending manual approval. This is
        the only case where the error body is meant to drive a *pending
        screen* rather than a form error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            dealer:
              value:
                error:
                  code: DEALER_PENDING_REVIEW
                  message: Your dealer account is awaiting support review — you'll get an email once it's
                    approved.
                  details:
                    role: dealer
                    status: pending
                  request_id: 9e8d7c6b-5a49-4e3d-8c2b-1a0f9e8d7c6b
            support:
              value:
                error:
                  code: SUPPORT_PENDING_REVIEW
                  message: Your support account is pending approval. Our team reviews every request —
                    you'll get an email once it's approved.
                  details:
                    role: support
                    status: pending
                  request_id: 2c3d4e5f-6a7b-4c8d-9e0f-1a2b3c4d5e6f
...
    conflict:
      description: |-
        The request is valid but the current state forbids it. `code`
        distinguishes: `CONFLICT` (generic, `details` lists the legal
        transitions), `ALREADY_EXISTS` (duplicate signup, second live quote
        on the same request), `INVALID_TRANSITION` (deal or quote status
        machine rejected the move), `TOKEN_ALREADY_USED` (refresh replay).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            illegal_deal_transition:
              value:
                error:
                  code: INVALID_TRANSITION
                  message: A matched deal cannot move straight to completed.
                  details:
                    current: matched
                    attempted: completed
                    allowed_transitions:
                    - negotiation
                    - cancelled
                  request_id: 5e6f7081-9a2b-4c3d-8e4f-5a6b7c8d9e0f
            duplicate_quote:
              value:
                error:
                  code: ALREADY_EXISTS
                  message: You already have a live quote on this request. Revise it instead.
                  details:
                    existing_quote_id: 724c2d61-b1be-47ba-9458-0f4001204a1b
                  request_id: 6f708192-0a3b-4d5e-9f6a-7b8c9d0e1f2a
    validation_error:
      description: |-
        Body, path or query failed validation. `details` is a
        field → message map so the client can highlight the offending input
        rather than showing a generic banner.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            field_errors:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Some fields need attention.
                  details:
                    email:
                    - That email is already registered.
                    budget_max:
                    - budget_max must be greater than or equal to budget_min.
                  request_id: 0718293a-4b5c-4d6e-8f70-9a1b2c3d4e5f
    invalid_token:
      description: Expired, already-used, or malformed one-shot token (reset link, WS ticket, chat id)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            expired_ticket:
              value:
                error:
                  code: TOKEN_EXPIRED
                  message: That link has expired. Request a new one.
                  details: null
                  request_id: 18293a4b-5c6d-4e7f-8071-9a2b3c4d5e6f
    rate_limited:
      description: |-
        Rate limit hit. `details` carries `retry_after` (seconds) and the
        limit name, so the client can show a countdown instead of a dead
        error state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            ai_rate:
              value:
                error:
                  code: RATE_LIMITED
                  message: You're sending a lot of questions. Give it a moment.
                  details:
                    limit: ai:20/min
                    retry_after: 42
                  request_id: 293a4b5c-6d7e-4f80-8192-9a3b4c5d6e7f
  schemas:
...
    BuyerSignupRequest:
      type: object
      required:
      - full_name
      - email
      - phone
      - password
      - terms_accepted
      properties:
        full_name:
          type: string
          minLength: 1
          maxLength: 120
          example: Adithyaa Anand
        email:
          type: string
          format: email
          maxLength: 320
          example: adithyaa.ma@gmail.com
        phone:
          type: string
          pattern: ^\+?[1-9]\d{6,14}$
          example: '+14695550142'
        password:
          type: string
          format: password
          minLength: 6
          description: |-
            Hashed with Argon2id, never logged, never returned.

            ⚠️ The product copy says **6 characters minimum** because that
            was explicitly specified, but the reset-password route enforces
            8. This is a known inconsistency, not an oversight: 6 characters
            is not a safe floor. It is recorded rather than silently changed
            because changing the product copy was not authorised.
        terms_accepted:
          type: boolean
          description: |-
            Must be literally `true`; a false value is a 422
            `VALIDATION_ERROR` and no account row is created.

            The signup form's Terms checkbox maps directly to this field — it is
            the only client-supplied half of the consent record.
          example: true
        terms_version:
          type: string
          description: |-
            Advisory only. The server overwrites this with its own
            `TERMS_VERSION` constant — a client must not be able to claim
            acceptance of a different version than the one it was shown.
          example: '2026-09-01'
        terms_accepted_at:
          type: string
          format: date-time
          description: |-
            Server timestamp of acceptance. If the client sends a value it is
            ignored and overwritten — the server's clock is the record.

            `NULL` on a pre-existing account means consent is absent, not
            unknown, and that account must re-accept. It is never backfilled
            with the migration timestamp, which would fabricate consent.
    DealerSignupRequest:
      type: object
      required:
      - full_name
      - email
      - phone
      - password
      - dealership_name
      - branch_name
      - dealer_license
      - state_id
      - address
      - terms_accepted
      properties:
        full_name:
          type: string
          minLength: 1
          maxLength: 120
        email:
          type: string
          format: email
          maxLength: 320
        phone:
          type: string
          pattern: ^\+?[1-9]\d{6,14}$
        password:
          type: string
          format: password
          minLength: 6
          description: Argon2id. Same 6-vs-8 character caveat as buyer signup.
        dealership_name:
          type: string
          minLength: 1
          maxLength: 150
          example: Navee Motors
        branch_name:
          type: string
          minLength: 1
          maxLength: 120
          example: Frisco
        dealer_license:
          type: string
          minLength: 1
          maxLength: 60
          description: Dealership licence number, verified by support against the state registry.
          example: US030202
        state_id:
          type: string
          format: uuid
          description: Branch state. Drives the dealer feed's default geo centre.
        address:
          type: string
          minLength: 1
          maxLength: 300
          example: 4100 Legacy Dr, Frisco, TX
        website:
          type: string
          format: uri
          maxLength: 300
          nullable: true
        brand_ids:
          type: array
          description: Brands this branch carries. Feeds the feed's brand filter; no FK on the array elements
            (Postgres cannot), so orphans are possible and ignored.
          items:
            type: string
            format: uuid
        terms_accepted:
          type: boolean
          enum:
          - true
          description: Must be literally `true` — the Terms checkbox is not optional. A 422 `VALIDATION_ERROR`
            is returned and no row is created otherwise.
        terms_version:
          type: string
          description: Advisory only; overwritten by the server's `TERMS_VERSION`.
        terms_accepted_at:
          type: string
          format: date-time
          description: Server timestamp; a client-supplied value is ignored. Written at the same moment
            as the `users` row, so an account never exists without its consent record.
    SupportSignupRequest:
      type: object
      required:
      - full_name
      - email
      - phone
      - password
      - address
      - state_id
      - terms_accepted
      description: |-
        Support team accounts require manual approval.

        Field names here are the exact signup form fields: Full Name, Phone
        Number, Location Address (split into `address` + `state_id`), optional
        Extra information, Email address, and password.
      properties:
        full_name:
          type: string
          minLength: 1
          maxLength: 120
          example: Support03
        phone:
          type: string
          pattern: ^\+?[1-9]\d{6,14}$
          example: '+14695550144'
        address:
          type: string
          minLength: 1
          maxLength: 300
          description: Location Address — street, city, postal code
          example: 2200 Ross Ave, Dallas, TX
        state_id:
          type: string
          format: uuid
          description: Location Address — state, from GET /reference/states
        extra_info:
          type: string
          maxLength: 2000
          description: Extra information (optional) — free text for the reviewing agent
          example: Former dealership service manager, 6 years.
          nullable: true
        email:
          type: string
          format: email
          maxLength: 320
          example: support03@gmail.com
        password:
          type: string
          format: password
          minLength: 6
          description: Argon2id. Same 6-vs-8 character caveat as buyer signup.
        terms_accepted:
          type: boolean
          enum:
          - true
          description: |-
            Must be literally `true` — the Terms checkbox is not optional for agents
            any more than it is for buyers or dealers. A 422 `VALIDATION_ERROR` is
            returned and no row is created.
          example: true
        terms_version:
          type: string
          description: |-
            Advisory only; overwritten by the server's `TERMS_VERSION`. A support agent
            handles other people's data, so the consent record is worth storing even
            though the account starts inactive pending approval.
          example: '2026-09-01'
        terms_accepted_at:
          type: string
          format: date-time
          description: |-
            Server timestamp; a client-supplied value is ignored. Written at the same
            moment as the `users` row, so no account ever exists without its consent
            record.
          example: '2026-09-01T14:22:07Z'
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    TokenPair:
      type: object
      required:
      - access_token
      - refresh_token
      - token_type
      - expires_in
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
          description: Single-use. Presenting it again revokes the whole family.
        token_type:
          type: string
          enum:
          - bearer
        expires_in:
          type: integer
          description: Access token lifetime in seconds
          example: 1800
    AuthSession:
      allOf:
      - $ref: '#/components/schemas/TokenPair'
      - type: object
        required:
        - profile
        properties:
          profile:
            $ref: '#/components/schemas/Profile'
    PendingApproval:
      type: object
      description: |-
        Returned by dealer and support signup instead of tokens. The account
        row exists and is inactive; nothing about the session is usable yet.
      required:
      - pending
      - role
      - email
      - verification_id
      - ticket_id
      - message
      properties:
        pending:
          type: boolean
          enum:
          - true
        role:
          type: string
          enum:
          - dealer
          - support
        email:
          type: string
          format: email
        verification_id:
          type: string
          format: uuid
          description: The `support_verifications` row an agent will act on.
        ticket_id:
          type: string
          description: Human reference — `DV<epoch>` for dealers, `SA<epoch>` for agents.
          example: DV1788793917
        message:
          type: string
          description: |-
            The exact copy the pending screen shows. Dealers and agents get
            different sentences, not one templated string.
...
    MeResponse:
      allOf:
      - $ref: '#/components/schemas/Profile'
      - type: object
        description: |-
          The client routes on `role` — there is no role picker anywhere in
          the UI, and a dealer pending approval renders a pending state
          rather than a nav.
        properties:
          preferences:
            description: Present for buyers with a stored preference row, else null.
            allOf:
            - $ref: '#/components/schemas/BuyerPreference'
            nullable: true
...
  examples:
    BuyerProfile:
      summary: The seeded buyer
      value:
        id: bdc31c9e-f6ed-40d0-a32f-706062f81d77
        role: buyer
        full_name: Adithyaa Anand
        email: adithyaa.ma@gmail.com
        phone: '+14695550142'
        avatar_url: https://cdn.drivedeal.example/avatars/adithyaa.png
        is_active: true
        verification_status: null
        verified_at: null
        created_at: '2026-07-14 10:02:11+00:00'
    MeBuyer:
      summary: /auth/me for a buyer with a stored preference
      value:
        id: bdc31c9e-f6ed-40d0-a32f-706062f81d77
        role: buyer
        full_name: Adithyaa Anand
        email: adithyaa.ma@gmail.com
        phone: '+14695550142'
        avatar_url: https://cdn.drivedeal.example/avatars/adithyaa.png
        is_active: true
        verification_status: null
        verified_at: null
        created_at: '2026-07-14 10:02:11+00:00'
        preferences:
          profile_id: bdc31c9e-f6ed-40d0-a32f-706062f81d77
          brand_id: b5a5e627-7f0f-4d6f-a7e5-1b2c3d4e5f60
          other_brand_ids:
          - c6d7e8f9-0a1b-4c2d-8e3f-4a5b6c7d8e9f
          model_preference: Bronco
          body_type: SUV
          seater_count: 7
          transmission: Automatic
          drivetrain: AWD
          fuel_type: Petrol
          condition: Used
          exterior_color: Blue
          min_year: 2021
          max_mileage: 60000
          budget_min: 40000
          budget_max: 65000
          must_have_features:
          - sunroof
          - leather
          - apple_carplay
          never_want_features:
          - manual_transmission
          source: ai_inferred
          confidence: 0.92
          created_at: '2026-09-01 09:18:20+00:00'
          updated_at: '2026-09-01 09:18:20+00:00'
        terms_accepted_at: '2026-09-01T14:22:07Z'
...